Check availability of 'path' before accessing it
The path for either $base or $rel might be empty. For example:

$base = "https://beesbuzz.biz"

PHP generates an error when attempting to access $pbase['path']
because it isn't set in this case (same applies to $prel)

A minor adjustment is to access the first character of the path using
'substr($prel['path'], 0, 1)' instead of '$prel['path'][0]' because
'$prel['path'][0]' returns an error if 'path' is null, where substr()
returns false!